fix(admin): resolve seed-admin script path resolution and dependency issues
- Move admin seed script to correct location (project root scripts/)
- Fix bcrypt dependency import using createRequire from family-tree node_modules
- Add comprehensive input validation and error handling
- Update documentation with correct usage instructions
- Remove incorrectly placed script from family-tree/scripts/
- Ensure cross-platform compatibility for Windows, macOS, Linux

Resolves: E13-T2 Admin Script Path Resolution (P1-CRITICAL)
