+++ family-tree/data/users.json
@@ -3,7 +3,7 @@
     {
       "id": "user-1",
       "username": "admin",
-      "password": "$2b$10$XdRXcGJO.GiZ0B.7VdPDAu2nVKwevd3yTKp.jQoQAdaW3d4a/Lhja",
+      "password": "$2b$10$/.RF46sliWhT18W8sqXLOOzSL1rS3HiwYreEugnpDt7HZqjqMIsfu",
       "role": "editor",
       "createdAt": "2024-07-30T00:00:00Z",
       "lastLogin": "2024-07-30T00:00:00Z"
